update dtos format

+++ src/main/java/com/orderingsystem/orderingsystem/entity/Bills.java
@@ -7,17 +7,18 @@ import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Table(
-        name = "bills",
-        indexes = {
-                @Index(name = "idx_bill_user", columnList = "user_id"),
-                @Index(name = "idx_bill_table", columnList = "table_id")
-        }
+    name = "bills",
+    indexes = {
+        @Index(name = "idx_bill_user", columnList = "user_id"),
+        @Index(name = "idx_bill_table", columnList = "table_id")
+    }
 )
 @Data
 @NoArgsConstructor
@@ -30,7 +31,7 @@ public class Bills {
     @NotNull
     @Min(0)
     @Column(nullable = false)
-    private Float total;
+    private BigDecimal total;
 
     @CreationTimestamp
     @Column(nullable = false, updatable = false)
