+++ src/main/java/com/orderingsystem/orderingsystem/entity/Products.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,8 +14,8 @@ import java.util.List;
 @Table(
     name = "products",
     indexes = {
-            @Index(name = "idx_product_name", columnList = "name"),
-            @Index(name = "idx_product_category", columnList = "category_id")
+        @Index(name = "idx_product_name", columnList = "name"),
+        @Index(name = "idx_product_category", columnList = "category_id")
     }
 )
 @Data
@@ -40,12 +41,12 @@ public class Products {
     @NotNull
     @Min(0)
     @Column(nullable = false)
-    private Float importPrice;
+    private BigDecimal importPrice;
 
     @NotNull
     @Min(0)
     @Column(nullable = false)
-    private Float salePrice;
+    private BigDecimal salePrice;
 
     @NotNull
     @Min(0)
